Work towards v3.3-b3
Added support for global keyword and for the global OM namespace
Added RRF interfaces for global, var, and set keyword
Object model lists may be queried successively
DCS falls back to 0:/sys if required firmware files could not be found in 0:/firmware
Added new comment tags for Ideamaker
Bug fix: M0/M1 didn't cancel already queued codes
Bug fix: Incorrect delta tilt corrections were saved to config-override.g
Bug fix: New seqs.volChanges[] field caused JSON errors
Bug fix: AppArmor template didn't allow processes to use mmap (which is required by .NET)

+++ src/DuetAPI/ObjectModel/Move/MoveQueueItem.cs
@@ -6,9 +6,9 @@
     public sealed class MoveQueueItem : ModelObject
     {
         /// <summary>
-        /// The minimum idle time in milliseconds, before we should start a move
+        /// The minimum idle time before we should start a move (in s)
         /// </summary>
-        public uint GracePeriod { get; set; }
+        public float GracePeriod { get; set; }
 
         /// <summary>
         /// Maximum number of moves that can be accomodated in the DDA ring
